Add compare_ratings

+++ Application/Ranking/CorrelationStudyProcessor.cpp
@@ -462,7 +462,8 @@ void CorrelationStudyProcessor::DrawCorrsComparison(const shared_ptr<Correlation
         EFT_PROF_DEBUG("Idx for: {:60} ==> {:3} & {:3}", np_name, idx_1, idx_2);
         if (idx_2 != -1) {
             if (settings->weighted) {
-                auto weight = static_cast<float>(idx_2 - idx_1) / nb_bins;
+                auto weight = static_cast<float>((int) idx_2 - (int) idx_1) / nb_bins;
+                //auto weight = static_cast<float>((idx_2 - idx_1)) / nb_bins;
                 EFT_PROF_DEBUG("weight: {:4} - {:4} / {:4} = {:4}", idx_2, idx_1, nb_bins, weight);
                 h->Fill(idx_1 + 1, idx_2 + 1, weight);
             }
